refactor: created a new hook to perform bid actions
